DRTII-269 Restrict earliest zip filename to now minus expiry millis  - add tests around each case

+++ server/src/main/scala/test/TestActors.scala
@@ -56,7 +56,7 @@ object TestActors {
     override def receiveCommand: Receive = reset orElse super.receiveCommand
   }
 
-  case class TestVoyageManifestsActor(override val now: () => SDateLike, expireAfterMillis: Long, snapshotInterval: Int)
+  case class TestVoyageManifestsActor(override val now: () => SDateLike, expireAfterMillis: Int, snapshotInterval: Int)
     extends VoyageManifestsActor(oneMegaByte, now, expireAfterMillis, Option(snapshotInterval)) {
 
     def reset: Receive = {
